#426 refactor: hoist browser fixtures to conftest for sibling test files
Pure move of browser_db, seeded_ids, live_server, browser, page and the
_free_port helper from test_a11y_browser.py to tests/api/admin/conftest.py so
the #367/#368 follow-up browser-test files can share them. Playwright stays a
lazy import (pytest.importorskip inside the browser fixture), so the conftest
imported by every admin test gains no browser-extra requirement; the sweep
module keeps its module-level importorskip for a clean whole-module skip.
Browser-only helpers (axe source, QUERY_PARAMS use, floor guard) stay in the
test module. No behavior change.
